feat(SPV-1429): move to utils
